laravel ok

+++ laravel/app/Http/Controllers/UserController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use Illuminate\Http\Request;
 use App\Models\User;
 use Illuminate\Support\Facades\Hash;
+use Tymon\JWTAuth\Facades\JWTAuth;
 
 class UserController extends Controller
 {
@@ -41,10 +42,14 @@ class UserController extends Controller
             'password' => Hash::make($request->password),
         ]);
 
-        $token = $user->createToken($request->name);
+        //$token = $user->createToken($request->name);
+
+        $token = JWTAuth::fromUser($user);
+
         return response()->json([
             'user' => $user,
-            'token' =>$token->plainTextToken,
+            //'token' =>$token->plainTextToken,
+            'token' =>$token,
              201
         ]);
     }
@@ -52,10 +57,24 @@ class UserController extends Controller
     /**
      * Display the specified resource.
      */
-    public function show(User $user)
+    /*public function show(User $user)
     {
         return response()->json($user, 200);
+    }*/
+    public function show($id)
+    {
+        $user = User::find($id);
+        if (!$user) {
+            return response()->json(['message' => 'Utilisateur non trouvé'], 404);
+        }
+
+        if (auth()->id() !== $user->id) {
+            return response()->json(['message' => 'Non autorisé'], 403);
+        }
+
+        return response()->json($user, 200);
     }
+
 
     /**
      * Update the specified resource in storage.
@@ -66,6 +85,10 @@ class UserController extends Controller
 
         if (!$user) {
             return response()->json(['message' => 'Utilisateur non trouvé'], 404);
+        }
+
+        if (!auth()->check()) {
+            return response()->json(['message' => 'Non autorisé'], 401);
         }
 
         $request->validate([
@@ -92,20 +115,44 @@ class UserController extends Controller
     /**
      * Remove the specified resource from storage.
      */
-    public function destroy(User $user)
+    /*public function destroy(User $user)
     {
+        $user->delete();
+
+        return response()->json(['message' => 'Utilisateur supprimé'], 200);
+    }*/
+    public function destroy($id)
+    {
+        $user = User::find($id);
+
+        if (!$user) {
+            return response()->json(['message' => 'Utilisateur non trouvé'], 404);
+        }
+
+        if (auth()->id() !== $user->id) {
+            return response()->json(['message' => 'Non autorisé'], 403);
+        }
+
         $user->delete();
 
         return response()->json(['message' => 'Utilisateur supprimé'], 200);
     }
 
     public function login(Request $request){
-        $request->validate([
+        $credentials =$request->validate([
             'email' => 'required|email|exists:users',
             'password' => 'required|string',
         ]);
+        
 
-        $user = User::where('email', $request->email)->first();
+        if (!$token = JWTAuth::attempt($credentials)) {
+            return response()->json(['message' => 'Unauthorized'], 401);
+        }
+    
+        return response()->json(['token' => $token]);
+
+
+        /*$user = User::where('email', $request->email)->first();
 
         if (!$user || !Hash::check($request->password, $user->password)) {
             return response()->json(['message' => 'Identifiants incorrects'], 401);
@@ -116,12 +163,13 @@ class UserController extends Controller
             'user connecté' => $user,
             'token'=>$token->plainTextToken,
              201
-        ]);
+        ]);*/
     }
 
-    public function logout(Request $request)
+    public function logout(/*Request $request*/)
     {
-        $request->user()->tokens()->delete();
+        //$request->user()->tokens()->delete();
+        JWTAuth::invalidate(JWTAuth::getToken());
 
         return response()->json(['message' => 'Déconnexion réussie'], 200);
     }
